pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools_scm]

[project]
name = "blockassist"
dynamic = ["version"]
description = "Distributed experiments built using the MBAG multiagent environment (https://github.com/cassidylaidlaw/minecraft-building-assistance-game)."
readme = "README.md"
requires-python = ">=3.8,<3.11"
dependencies = [
  "aiofiles",
  "boto3",
  "datasets",
  "eth-account",
  "hydra-core",
  "mbag-gensyn[malmo,rllib] @ git+https://github.com/gensyn-ai/minecraft-building-assistance-game.git@96d57776c3595c7ec94d115d05074051bc015f6a",
  "numpy>=1.21,<1.22",
  "requests",
  "six",
  "pydantic",
  "web3",
]

[project.optional-dependencies]
dev = ["ruff", "isort", "pytest", "pytest-asyncio"]
